backend/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
python-dotenv==1.0.0
alembic==1.12.0
aiosqlite==0.19.0
sqlalchemy==2.0.22
email-validator==2.0.0
Pyjwt==2.8.0